requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
gitdb==4.0.2
GitPython==3.1.0
idna==2.9
multidict==4.7.5
python-gitlab==2.0.1
requests==2.23.0
smmap==3.0.1
urllib3==1.25.8
yarl==1.4.2
